app/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
PyPDF2
dotenv
mlflow
watchdog